chore(express):  alter index.html before starting express app
