+++ Dockerfile
@@ -7,6 +7,8 @@ WORKDIR /usr/src/app
 COPY package.json yarn.lock ./
 RUN yarn install --frozen-lockfile --check-files --production=true
 COPY www ./www
+COPY template ./template
+COPY env.js ./
 COPY app.js ./
 COPY ./entrypoint.sh /
 RUN chmod +x /entrypoint.sh
