+++ app.js
@@ -1,6 +1,23 @@
 const express = require('express')
 const morgan = require('morgan')
 const compression = require('compression')
+const fs = require('fs')
+const jsdom = require('jsdom')
+const useEnv = require('./env')
+
+fs.readFile('./template/index.html', 'utf8', (err, file) => {
+  if (err) {
+    throw err;
+  }
+  const { JSDOM } = jsdom
+  const dom = new JSDOM(file);
+  const html = useEnv(dom.window.document).documentElement.outerHTML;
+  fs.writeFile('./www/index.html', html, 'utf8', (err) => {
+    if (err) {
+      throw err;
+    }
+  });
+});
 
 const app = express();
 app.use(morgan('combined'));
