+++ entrypoint.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-originalfile="/usr/src/app/www/js/env.js"
+originalfile="/usr/src/app/env.js"
 tmpfile=$(mktemp)
 cp --attributes-only --preserve $originalfile $tmpfile
 cat $originalfile | envsubst | tee $tmpfile &&  mv $tmpfile $originalfile
