+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "compression": "^1.7.4",
     "express": "^4.17.1",
-    "morgan": "^1.10.0"
+    "morgan": "^1.10.0",
+    "jsdom": "^17.0.0"
   }
 }
